requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
openai
pinecone-client
transformers
torch
sentence-transformers
requests
pydantic
python-multipart
